requirements.txt
python-telegram-bot[ext]==21.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
aiohttp==3.11.11
